% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{across_operators}
\alias{across_operators}
<<<<<<< HEAD
\title{FUNCTION: across_operators}
\usage{
across_operators(data)
}
\description{
This function will be used to generate a sequence of values between 0.01 and 1.00
that will be used to test the mutation score of each fraction of reduced set of mutants
on a per-operator basis. Additionally, we will use this data to calculate mutation scores
and costs across all operators for each percentage.
=======
\title{FUNCTION: across operators}
\usage{
across_operators(d, i, j)
}
\description{
Perform random sampling across operators
>>>>>>> mrstudyr2/master
}

