% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{FUNCTION: analyse}
\usage{
analyse(data)
}
\description{
This function will encompass all of the reduction techniques, returning
a single dataframe with the data from performing all techniques.
}

