% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_calculations}
\alias{analyse_calculations}
\title{FUNCTION: analyse_calculations}
\usage{
analyse_calculations(data)
}
\description{
This function for each percent will be used to show the effectiveness
and efficiency of each approach.
}

