% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_correlation}
\alias{analyse_correlation}
\title{FUNCTION: analyse_correlation}
\usage{
analyse_correlation(d)
}
\description{
Calculate Kendall's tau_b correlation coefficient between the reduced and original mutation score
}

