% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_mutation_score}
\alias{analyse_mutation_score}
\title{FUNCTION: analyse_mutation_score}
\usage{
analyse_mutation_score(data)
}
\description{
This function will be used to calculate mutation scores. It will calculate
mutation scores for both the entire set of mutants and the selected
set of mutants. Then we will use this to compare the effectiveness of
each selective approach.
}

