% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_percents_average_reduced_mutation_score}
\alias{analyse_percents_average_reduced_mutation_score}
\title{FUNCTION: analyse_percents_average_reduced_mutation_score}
\usage{
analyse_percents_average_reduced_mutation_score(data)
}
\description{
This function will calculate the averages of all of the thirty trials and
display them in an easy-to-view data frame.
}

