% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_percents_error}
\alias{analyse_percents_error}
\title{FUNCTION: analyse_percents_error}
\usage{
analyse_percents_error(data)
}
\description{
This function will calculate both the mean absolute error
and rooth mean squared error for all percentages
for the random sampling data. This allows us to plot both against each
other in a visualization.
}

