% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_random_sampling}
\alias{analyse_random_sampling}
\title{FUNCTION: analyse_random_sampling}
\usage{
analyse_random_sampling(data, f = "fast")
}
\description{
This function will be used to run selective mutation per schema and collect
data in a single data frame
}

