% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_reduction}
\alias{analyse_reduction}
\title{FUNCTION: analyse_reduction}
\usage{
analyse_reduction(o, r)
}
\description{
This function calculates the percentage in reduction of creation cost
between the original and reduces sets. As input this function takes the
}

