% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_select_mutation_score}
\alias{analyse_select_mutation_score}
\title{FUNCTION: analyse_select_mutation_score}
\usage{
analyse_select_mutation_score(data)
}
\description{
This function will be used to generate a sequence of values between 0.01 and 1.00
that will be used to test the mutation score of each fraction of reduced set of mutants.
It returns a vector which we will calculate the average of for each fraction and then
append that avgeraged fraction to the entire mutant set.
}

