% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_total_time}
\alias{analyse_total_time}
\title{FUNCTION: analyse_total_time}
\usage{
analyse_total_time(data)
}
\description{
This function will be used to calculate the total time (in ms) of performing
mutation analysis with the given set of mutants.
}

