% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_across_operators}
\alias{analyze_across_operators}
\title{FUNCTION: analyze_across_operators}
\usage{
analyze_across_operators(d)
}
\description{
This function will perform sampling across all operators --- this is not operators selection as alll
operators will be present.
}

