% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_correlation}
\alias{analyze_correlation}
\title{FUNCTION: analyze_correlation}
\usage{
analyze_correlation(d)
}
\description{
Calculate Kendall's tau_b correlation coefficient between the reduced and original mutation score
}

