% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_incremental}
\alias{analyze_incremental}
\title{FUNCTION: analyze_incremental}
\usage{
analyze_incremental(d, partition_size = 1)
}
\description{
Analyze how reducing the set incrementally effects the error between MS and MS'
}

