% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_keep}
\alias{analyze_keep}
\title{FUNCTION: analyze_keep}
\usage{
analyze_keep(d)
}
\description{
Generate a neighborhood (keep_data). This will be used as a starting point from which to incrementally bit-flip.
}

