% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_reduced_rmse}
\alias{analyze_reduced_rmse}
\title{FUNCTION: analyze_reduced_rmse}
\usage{
analyze_reduced_rmse(d)
}
\description{
Calculate the rmse of the mutation score of the chosen set to the original set of mutants.
}

