% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_selective_mutation}
\alias{analyze_selective_mutation}
\title{FUNCTION: analyze_selective_mutation}
\usage{
analyze_selective_mutation(d, o)
}
\description{
This function will perform selective mutation over 30 trials
}

