% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_selective_random}
\alias{analyze_selective_random}
\title{FUNCTION: analyze_selective_random}
\usage{
analyze_selective_random(d, ops)
}
\description{
This function will analyze a select set of operators over percentages
}

