% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_selective_random_mutation}
\alias{analyze_selective_random_mutation}
\title{FUNCTION: analyze_selective_random_mutation}
\usage{
analyze_selective_random_mutation(d, o)
}
\description{
This function will analyze a select set of operators over percentages
}

