% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply_operator_model}
\alias{apply_operator_model}
\title{FUNCTION: apply_operator_model}
\usage{
apply_operator_model(d, model)
}
\description{
Apply the model generated from performing hill climbing to a schema.
}

