% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply_operator_model_per_schema}
\alias{apply_operator_model_per_schema}
\title{FUNCTION: apply_operator_model_per_schema}
\usage{
apply_operator_model_per_schema(d, model)
}
\description{
Apply the model generated from performing hill climbing to each schema.
}

