% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_effectiveness}
\alias{calculate_effectiveness}
\title{FUNCTION: calculate_effectiveness}
\usage{
calculate_effectiveness(d, p = FALSE)
}
\description{
Calculate the effectiveness of a reduction technique on a per-schema (optional, per-percentage) basis.
}

