% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_fractional_operator_costs}
\alias{calculate_fractional_operator_costs}
\title{FUNCTION: calculate_fractional_operator_costs}
\usage{
calculate_fractional_operator_costs(d)
}
\description{
Calculate the fractional costs of mutants per operator.
}

