% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_fractional_operator_frequencies}
\alias{calculate_fractional_operator_frequencies}
\title{FUNCTION: calculate_fractional_operator_frequencies}
\usage{
calculate_fractional_operator_frequencies(d)
}
\description{
Calculate the fractional frequencies of mutants per operator.
}

