% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_highest_correlation}
\alias{calculate_highest_correlation}
\title{FUNCTION: calculate_highest_correlation}
\usage{
calculate_highest_correlation(d)
}
\description{
Determine which step is has the highest correlation
}

