% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{calculate_mutation_score_correlation}
\alias{calculate_mutation_score_correlation}
\title{FUNCTION: calculate_mutation_score_correlation}
\usage{
calculate_mutation_score_correlation(data)
}
\description{
This function will calculate the correlation between percentages and the original
mutation score for that percent. This is only possible if we look at all schemas at the
same time for a given percent. Otherwise, the standard deviation is zero and correlation
is not able to be calculated
}

