% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_neighborhood_size}
\alias{calculate_neighborhood_size}
\title{FUNCTION: calculate_neighborhood_size}
\usage{
calculate_neighborhood_size(d)
}
\description{
This function extracts the size of the neighborhood from the data.
}

