% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_per_dbms_mutation_scores}
\alias{calculate_per_dbms_mutation_scores}
\title{FUNCTION: calculate_per_dbms_mutation_scores}
\usage{
calculate_per_dbms_mutation_scores(d)
}
\description{
In an attempt to explain why RS is so effective, calculate the mutation scores of data for each dbms.
}

