% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_per_schema_fractional_operator_frequencies}
\alias{calculate_per_schema_fractional_operator_frequencies}
\title{FUNCTION: calculate_per_schema_fractional_operator_frequencies}
\usage{
calculate_per_schema_fractional_operator_frequencies(d)
}
\description{
Calculate the fractional frequencies of mutants per operator on a per-schema basis.
}

