% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_per_trial_percentage_effectiveness}
\alias{calculate_per_trial_percentage_effectiveness}
\title{FUNCTION: calculate_per_trial_percentage_effectiveness}
\usage{
calculate_per_trial_percentage_effectiveness(d)
}
\description{
Calculate the effectiveness of a reduction technique on a per-trial, per-percentage basis.
}

