% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_percent_effectiveness}
\alias{calculate_percent_effectiveness}
\title{FUNCTION: calculate_percent_effectiveness}
\usage{
calculate_percent_effectiveness(d)
}
\description{
Calculate the effectiveness of a reduction technique on a per-schema per-percentage basis.
}

