% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_percent_summary}
\alias{calculate_percent_summary}
\title{FUNCTION: calculate_percent_summary}
\usage{
calculate_percent_summary(d)
}
\description{
Calculate the effectiveness of a reduction technique.
The metrics to determine effectiveness will be Kendall's tau_b correlation coefficient,
mae and rmse.
}

