% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_best_fit_data}
\alias{collect_best_fit_data}
\title{FUNCTION: collect_best_fit_data}
\usage{
collect_best_fit_data(d)
}
\description{
Filter the data where fitness equals the best fitness for a given schema.
}

