% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_best_keep_data}
\alias{collect_best_keep_data}
\title{FUNCTION: collect_best_keep_data}
\usage{
collect_best_keep_data(b, k)
}
\description{
Filter the collected keep data from all positions and filter where position was best fit.
}

