% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_best_step_data}
\alias{collect_best_step_data}
\title{FUNCTION: collect_best_step_data}
\usage{
collect_best_step_data(b, k)
}
\description{
Filter the collected keep data from all steps and filter where step had highest correlation.
}

