% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_chosen_percent_data}
\alias{collect_chosen_percent_data}
\title{FUNCTION: collect_chosen_percent_data}
\usage{
collect_chosen_percent_data(d, i)
}
\description{
Collect the data for a user-specified percentage
}

