% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_chosen_trial_data}
\alias{collect_chosen_trial_data}
\title{FUNCTION: collect_chosen_trial_data}
\usage{
collect_chosen_trial_data(d, t)
}
\description{
Collect the data for a user-specified trial
}

