% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_highest_correlation_data}
\alias{collect_highest_correlation_data}
\title{FUNCTION: collect_highest_correlation_data}
\usage{
collect_highest_correlation_data(d)
}
\description{
Filter the data where correlation equals highest correlation
}

