% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_ignore_data}
\alias{collect_ignore_data}
\title{FUNCTION: collect_ignore_data}
\usage{
collect_ignore_data(d)
}
\description{
Filter data to only include data that is 'ignored' i.e., a boolean value of 'FALSE' in the keep column.
}

