% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_keep_data}
\alias{collect_keep_data}
\title{FUNCTION: collect_keep_data}
\usage{
collect_keep_data(d)
}
\description{
Filter data to only include data that is 'kept' i.e., a boolean value of 'TRUE' in the keep column.
}

