% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_normal_data}
\alias{collect_normal_data}
\title{FUNCTION: collect_normal_data}
\usage{
collect_normal_data(d)
}
\description{
This function pulls all data with type equal to NORMAL from the original data.
}

