% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_schema_operator_data}
\alias{collect_schema_operator_data}
\title{FUNCTION: collect_schema_operator_data}
\usage{
collect_schema_operator_data(d)
}
\description{
Group data by dbms, schema and operator
}

