% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_schema_position_data}
\alias{collect_schema_position_data}
\title{FUNCTION: collect_schema_step_data <- function()}
\usage{
collect_schema_position_data(b, k)
}
\description{
Given the best-fit data, gather the keep data for the respective position
for a given schema.
}

