% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_study_schemas}
\alias{collect_study_schemas}
\title{FUNCTION: collect_study_schemas}
\usage{
collect_study_schemas(d)
}
\description{
This function will reduce the data to only analyze the schemas discussed
and presented in the accompanying tool paper.
(CoffeeOrders, Employee, Inventory, Iso3166, JWhoisServer, MozillaPermissions, NistWeather, Person, Products)
}

