% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_technique_trial_data}
\alias{collect_technique_trial_data}
\title{FUNCTION: collect_technique_trial_data}
\usage{
collect_technique_trial_data(d)
}
\description{
Group by technique and trial for summarization across schemas.
}

