% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_trial_operator_data}
\alias{collect_trial_operator_data}
\title{FUNCTION: collect_trial_operator_data}
\usage{
collect_trial_operator_data(d)
}
\description{
Group data by dbms, trial and operator (not schema because this is across-schema) for generating a generalized model.
}

