% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_technique_group_head_to_head}
\alias{compare_technique_group_head_to_head}
\title{FUNCTION: compare_technique_group_head_to_head}
\usage{
compare_technique_group_head_to_head(d, a, b)
}
\description{
This function is responsible to comparing two technique groups head-to-head.
}

