% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_effect_size_test}
\alias{create_effect_size_test}
\title{FUNCTION: create_effect_size_test}
\usage{
create_effect_size_test(d)
}
\description{
Create the graphs associated with the pairwise effect size calculations.
}

