% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_mutation_score_graphs}
\alias{create_mutation_score_graphs}
\title{FUNCTION: create_mutation_score_graphs}
\usage{
create_mutation_score_graphs()
}
\description{
This function will call the necessary function to create the mutation score graphs
}

