% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_pairwise_wilcoxon_rank_sum_test}
\alias{create_pairwise_wilcoxon_rank_sum_test}
\title{FUNCTION: create_pairwise_wilcoxon_rank_sum_test}
\usage{
create_pairwise_wilcoxon_rank_sum_test(d)
}
\description{
Create the graphs associated with the pairwise Wilcoxon ranked-sum test.
The input to this function is the combined technique data.
}

