% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{create_results_tables_per_schema}
\alias{create_results_tables_per_schema}
\title{FUNCTION: create_results_tables_per_schema}
\usage{
create_results_tables_per_schema(d)
}
\description{
Create LaTeX tables of data for each schema.
}

