% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{cumulative_sum_time}
\alias{cumulative_sum_time}
\title{FUNCTION: cumulative_sum_time}
\usage{
cumulative_sum_time(data)
}
\description{
This function calculates the sum of times for each group---in this case, each mutant of
the same type, of the same operator, in the same schema. This will allow us to analyse
the overall cost per mutant type.
}

