% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effectsize_accurate}
\alias{effectsize_accurate}
\title{FUNCTION: effectsize_accurate}
\usage{
effectsize_accurate(d, f)
}
\description{
Calculate the Vargha-Delaney Effect Size using the approach that is claimed to avoid accuracy errors
Author: @gkapfham https://github.com/gkapfham/virtualmutationanalysis
}

