% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effectsize_calculate_accurate}
\alias{effectsize_calculate_accurate}
\title{FUNCTION: effectsize_calculate_accurate}
\usage{
effectsize_calculate_accurate(r1, m, n)
}
\description{
Calculate the actual Vargha-Delaney A value using the more accurate approach.
Author: @gkapfham https://github.com/gkapfham/virtualmutationanalysis
}

