% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effectsize_calculate_standard}
\alias{effectsize_calculate_standard}
\title{FUNCTION: effectsize_calculate_standard}
\usage{
effectsize_calculate_standard(r1, m, n)
}
\description{
Calculate the actual Vargha-Delaney A value using the standard approach.
Author: @gkapfham https://github.com/gkapfham/virtualmutationanalysis
}

