% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effectsize_default}
\alias{effectsize_default}
\title{FUNCTION: effectsize_default}
\usage{
effectsize_default(d, f, ac)
}
\description{
Default way to calculate the Vargha-Delaney Effect Size using the approach closest to published SBSE papers.
Author: @gkapfham https://github.com/gkapfham/virtualmutationanalysis
}

