% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effectsize_interpret}
\alias{effectsize_interpret}
\title{FUNCTION: effectsize_interpret}
\usage{
effectsize_interpret(a)
}
\description{
Interpret the Vargha-Delaney Effect Size using the approach proposed by Chris Wright.
Author: @gkapfham https://github.com/gkapfham/virtualmutationanalysis
}

