% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effectsize_standard}
\alias{effectsize_standard}
\title{FUNCTION: effectsize_standard}
\usage{
effectsize_standard(d, f)
}
\description{
Calculate the Vargha-Delaney Effect Size using the approach closest to published SBSE papers.
Author: @gkapfham https://github.com/gkapfham/virtualmutationanalysis
}

