% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effectsize_wright}
\alias{effectsize_wright}
\title{FUNCTION: effectsize_wright}
\usage{
effectsize_wright(sample1, sample2)
}
\description{
Calculate the Vargha-Delaney Effect Size using the approach proposed by Chris Wright.
Author: @gkapfham https://github.com/gkapfham/virtualmutationanalysis
}

