% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_incremental}
\alias{evaluate_incremental}
\title{FUNCTION: evaluate_incremental}
\usage{
evaluate_incremental(d, step_number)
}
\description{
Using keep column, calculate the original and reduced sets mutation score, error and execution time.
}

