% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_original_data_per_dbms}
\alias{evaluate_original_data_per_dbms}
\title{FUNCTION: evaluate_original_data_per_dbms}
\usage{
evaluate_original_data_per_dbms(d)
}
\description{
Evaluate the data prior to performing a reduction technique.
}

