% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_original_data_per_schema}
\alias{evaluate_original_data_per_schema}
\title{FUNCTION: evaluate_original_data_per_schema}
\usage{
evaluate_original_data_per_schema(d)
}
\description{
Evaluate the data prior to performing a reduction technique on a per schema basis.
}

