% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_random_sampling}
\alias{evaluate_random_sampling}
\title{FUNCTION: evaluate_random_sampling}
\usage{
evaluate_random_sampling(d, i, j)
}
\description{
Evaluate the effectiveness of random sampling at a given percent threshold.
}

