% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_reduction_technique}
\alias{evaluate_reduction_technique}
\title{FUNCTION: evaluate_reduction_technique}
\usage{
evaluate_reduction_technique(o, r)
}
\description{
Evaluate the effectiveness of a reduction technique given the original (pre-reduction data) and the reduced data
}

