% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_reduction_technique_across}
\alias{evaluate_reduction_technique_across}
\title{FUNCTION: evaluate_reduction_technique_across}
\usage{
evaluate_reduction_technique_across(o, r, s)
}
\description{
Evaluate the effectiveness of a reduction technique given the original (pre-reduction data) and the reduced data
}

