% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_selective_random}
\alias{evaluate_selective_random}
\title{FUNCTION: evaluate_selective_random}
\usage{
evaluate_selective_random(d, o, i, j)
}
\description{
Evaluate the effectiveness of selective random at a given percent threshold (i), for a given set of operators, (o).
}

