% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{facet_visualise_mutation_score_across_schemas}
\alias{facet_visualise_mutation_score_across_schemas}
\title{FUNCTION: facet_visualise_mutation_score_across_schemas}
\usage{
facet_visualise_mutation_score_across_schemas(data, a, b, c, d)
}
\description{
Produces a facetted visualisations of the reduced mutation scores
across all schemas at four distinct percentages.
}

