% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{generate_keep}
\alias{generate_keep}
\title{FUNCTION: generate_keep}
\usage{
generate_keep(d, trial)
}
\description{
This function generates a data frame consisting of a starting neighborhood that has been randomly-generated.
}

