% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_operator_model}
\alias{generate_operator_model}
\title{FUNCTION: generate_operator_model}
\usage{
generate_operator_model(d)
}
\description{
This model describes which operators should be kept versus removed to stay strongly correlated
while reducing the mutant set.
}

