% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{group_mutation_data}
\alias{group_mutation_data}
\title{FUNCTION: group_mutation_data}
\usage{
group_mutation_data(data)
}
\description{
We want to group the mutation data by schema, operator and mutant.
This will allow us to compare timings of similar mutants and then
provide sums of all timings per group.
}

