% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_apply_operator_model}
\alias{helper_apply_operator_model}
\title{FUNCTION: helper_apply_operator_model}
\usage{
helper_apply_operator_model(d, model)
}
\description{
FUNCTION: helper_apply_operator_model
}

