% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_bitflip_keep_across}
\alias{helper_bitflip_keep_across}
\title{FUNCTION: helper_bitflip_keep_across}
\usage{
helper_bitflip_keep_across(d)
}
\description{
Currently, this function negates boolean values i.e., TRUE -> FALSE, FALSE -> TRUE. Position is
used to idicate the current position to bitflip and step_size is the number of subsequent of mutants.
**We could add another parameter 'group_by' so that instead of just
flipping consecutive values, we could flip based on some group.
}

