% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_first_flip}
\alias{helper_first_flip}
\title{FUNCTION: helper_first_flip}
\usage{
helper_first_flip(d, s)
}
\description{
This function is responsible for doing the first flip. This includes choosing the random start
positions and ignoring the respective mutants.
}

