% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_flip}
\alias{helper_flip}
\title{FUNCTION: helper_flip}
\usage{
helper_flip(d)
}
\description{
This function is responsible for doing the first flip. This includes choosing the random start
positions and ignoring the respective mutants.
}

