% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_gather_keep_data}
\alias{helper_gather_keep_data}
\title{FUNCTION: helper_gather_keep_data}
\usage{
helper_gather_keep_data(b, k)
}
\description{
Gather data only the keep data cooresponding the best fit step for a given schema.
}

