% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_select_operator_data}
\alias{helper_select_operator_data}
\title{FUNCTION: helper_select_operator_data}
\usage{
helper_select_operator_data(d, model)
}
\description{
This function is responsible for selecting the appropriate amount of mutants for an operator
according to the generated model.
}

