% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_hill_climbing_size_data}
\alias{join_hill_climbing_size_data}
\title{FUNCTION: join_hill_climbing_size_data}
\usage{
join_hill_climbing_size_data(s, m, l, step_size_small, step_size_medium,
  step_size_large)
}
\description{
Join the data collected from different step sizes of the hill climbing reduction technique.
}

