% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_reduced_original_numerator_denominator_time_data}
\alias{join_reduced_original_numerator_denominator_time_data}
\title{FUNCTION: join_reduced_original_numerator_denominator_time_data}
\usage{
join_reduced_original_numerator_denominator_time_data(rn, rd, on, od, rt, ot)
}
\description{
DESCRIPTION: Join together six tables containing the numerator and denominator counts as well as the mutation analysis time for the reduced and original mutant sets
}

