% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_schema_operator_time}
\alias{join_schema_operator_time}
\title{FUNCTION: join_schema_operator_time}
\usage{
join_schema_operator_time(a, b)
}
\description{
Join summarized data containing per-schema original time and per-operator times.
}

