% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_summarized_keep_ignore_data}
\alias{join_summarized_keep_ignore_data}
\title{FUNCTION: join_summarized_keep_ignore_data}
\usage{
join_summarized_keep_ignore_data(a, b)
}
\description{
Join the summarized, model data for percentage of mutants kept and ignored per-operator.
}

