% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_technique_data}
\alias{join_technique_data}
\title{FUNCTION: join_technique_data}
\usage{
join_technique_data(da, db, dc)
}
\description{
Join together the data collected from each experiment campaign for each reduction technique.
}

