% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_total_keep_ignore_data}
\alias{join_total_keep_ignore_data}
\title{FUNCTION: join_total_keep_ignore_data}
\usage{
join_total_keep_ignore_data(tc, kc, ic)
}
\description{
Join together the total, keep and ignore mutant data after performing across-schema hill-climbing.
}

