% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{perform_effectsize_accurate}
\alias{perform_effectsize_accurate}
\title{FUNCTION: perform_effectsize_accurate}
\usage{
perform_effectsize_accurate(d, m)
}
\description{
Calculate the pairwise effect size by comparing each technique.
}

