% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{perform_pairwise_wilcoxon_rank_sum_test}
\alias{perform_pairwise_wilcoxon_rank_sum_test}
\title{FUNCTION: perform_pairwise_wilcoxon_rank_sum_test}
\usage{
perform_pairwise_wilcoxon_rank_sum_test(d, m)
}
\description{
Perform a statistical analysis of the correlation for each mutant reduction technique.
}

