% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{perform_wilcoxon_accurate}
\alias{perform_wilcoxon_accurate}
\title{FUNCTION: perform_wilcoxon_accurate}
\usage{
perform_wilcoxon_accurate(d, m)
}
\description{
Calculate the wilcoxon ranked sum for all pairs of techniques
}

