% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data}
\alias{read_data}
\title{FUNCTION: read_data}
\usage{
read_data(dat)
}
\description{
Read in a data file containing all of the mutation analysis data. This is
the file containing all of the mutation scores and timings for all schemas
and databases.
}

