% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hypersql_avmdefaults}
\alias{read_hypersql_avmdefaults}
\title{FUNCTION: read_hypersql_avmdefault}
\usage{
read_hypersql_avmdefaults(dat)
}
\description{
Read in a data file containing all of the mutation analysis data for the hypersql dbms.
}

