% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_postgres_avmdefaults}
\alias{read_postgres_avmdefaults}
\title{FUNCTION: read_postgres_avmdefault}
\usage{
read_postgres_avmdefaults(dat)
}
\description{
Read in a data file containing all of the mutation analysis data for the postgres dbms.
}

