% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sqlite_avmdefaults}
\alias{read_sqlite_avmdefaults}
\title{FUNCTION: read_sqlite_avmdefaults}
\usage{
read_sqlite_avmdefaults(dat)
}
\description{
Read in a data file containing all of the mutation analysis data for sqlite avmdefaults. This is
the file containing all of the mutation scores and timings for all schemas
and databases.
}

