% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{reduce_incremental}
\alias{reduce_incremental}
\title{FUNCTION: reduce_incremental}
\usage{
reduce_incremental(d, step_number)
}
\description{
Using keep column, calculate the original and reduced sets mutation score, error and execution time.
}

