% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{reduce_keep}
\alias{reduce_keep}
\title{FUNCTION: reduce_keep}
\usage{
reduce_keep(d, trial)
}
\description{
Using keep column, calculate the original and reduced sets' mutation score, error and execution time.
}

