% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_current_best_correlation}
\alias{select_current_best_correlation}
\title{FUNCTION: select_current_best_correlation}
\usage{
select_current_best_correlation(d)
}
\description{
Get the highest correlation from a list of correlation values and return a single value
}

