% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_current_position}
\alias{select_current_position}
\title{FUNCTION: select_current_position}
\usage{
select_current_position(d)
}
\description{
Get the current position (a single value), not a column
}

