% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_current_step_size}
\alias{select_current_step_size}
\title{FUNCTION: select_current_step_size}
\usage{
select_current_step_size(d)
}
\description{
Get the current step size (a single value), not a column
}

