% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_distinct_dbms_schema}
\alias{select_distinct_dbms_schema}
\title{FUNCTION: select_distinct_dbms_schema}
\usage{
select_distinct_dbms_schema(d)
}
\description{
Select only unique combinations of dbms and schema from sample data
}

