% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_empirical_study_schemas}
\alias{select_empirical_study_schemas}
\title{FUNCTION: select_empirical_study_schemas}
\usage{
select_empirical_study_schemas(d)
}
\description{
This function will reduce the data to only analyse the schemas discussed
and presented in the accompanying tool paper.
(CoffeeOrders, Employee, Inventory, Iso3166, JWhoisServer, MozillaPermissions, NistWeather, Person, Products)
}

