% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_head_to_head_technique_groups}
\alias{select_head_to_head_technique_groups}
\title{FUNCTION: select_head_to_head_technique_groups}
\usage{
select_head_to_head_technique_groups(d, a, b)
}
\description{
Get only the data for two technique groups (e.g., HC and RS)
}

