% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_individual_operator_data}
\alias{select_individual_operator_data}
\title{FUNCTION: select_individual_operator_data}
\usage{
select_individual_operator_data(data, op)
}
\description{
This function will create a data frame containing only the data
for a specific operator. This function will be used in combination
with other functions to test the effectiveness of random sampling over
mutation operators.
}

