% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_individual_percent_data}
\alias{select_individual_percent_data}
\title{FUNCTION: select_individual_percent_data}
\usage{
select_individual_percent_data(data, k)
}
\description{
This function will create a data frame containing only the data
for a specific percentage. This function will be used in combination
with select_indiv_schema_data to overall create a data frame for each schema
and for each set of percentages. We will then use this to find the average of the
trials
}

