% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_individual_schema_data}
\alias{select_individual_schema_data}
\title{FUNCTION: select_individual_schema_data}
\usage{
select_individual_schema_data(data, s)
}
\description{
This function will create a data frame containing only the data
for a specific schema. The schema name will be passed as input
and the function will return only data for that particular schema.
}

