% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_k_percent}
\alias{select_k_percent}
\title{FUNCTION: select_k_percent}
\usage{
select_k_percent(data, k)
}
\description{
This function will be used to look at k% of all data.
This is referred to as uniform random sampling.
}

