% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_k_percent_per_operator}
\alias{select_k_percent_per_operator}
\title{FUNCTION: select_k_percent_per_operator}
\usage{
select_k_percent_per_operator(data, k, op)
}
\description{
This function will be used to look at k% of each operator.
This is referred to as stratified random sampling over mutation operators.
}

