% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_normal_data}
\alias{select_normal_data}
\title{FUNCTION: select_normal_data}
\usage{
select_normal_data(data)
}
\description{
This function pulls all data with type equal to NORMAL from the original
data provided.
}

