% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_operators}
\alias{select_operators}
\title{FUNCTION: select_operators}
\usage{
select_operators(d, o)
}
\description{
This function will be used to look at a select set of operators.
This is referred to as selective mutation
}

