% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_percentage_across_operators}
\alias{select_percentage_across_operators}
\title{FUNCTION: select_percentage_across_operators}
\usage{
select_percentage_across_operators(data, i)
}
\description{
This function will be used to generate a sequence of values between 0.01 and 1.00
that will be used to test the mutation score of each fraction of reduced set of mutants
on a per-operator basis. Additionally, we will use this data to calculate mutation scores
and costs across all operators for each percentage.
}

