% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_random_start_position}
\alias{select_random_start_position}
\title{FUNCTION: select_random_start_position}
\usage{
select_random_start_position(x)
}
\description{
For the hill climbing approach a start position needs to be randomly chosen.
This function randomly selects a start position between 1 and the number of observations associated
with a schema (i.e., the number of rows or number of mutants).
}

