% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_random_start_position_frac}
\alias{select_random_start_position_frac}
\title{FUNCTION: select_random_start_position_frac}
\usage{
select_random_start_position_frac(d)
}
\description{
Might merge into other similar function, still thinking
}

