% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_unique_methods}
\alias{select_unique_methods}
\title{FUNCTION: select_unique_methods}
\usage{
select_unique_methods(data)
}
\description{
This function returns all unique values in the "method" column
}

