% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_unique_operators}
\alias{select_unique_operators}
\title{FUNCTION: select_unique_operators}
\usage{
select_unique_operators(data)
}
\description{
This function returns all unique values in the "operator" column
}

