% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_unique_percentages}
\alias{select_unique_percentages}
\title{FUNCTION: select_unique_percentages}
\usage{
select_unique_percentages(data)
}
\description{
This function will create a data frame of all of the unique percentages.
It should return 1-100. I created this function because passing a sequence
to a function was not working
}

