% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_unique_schemas}
\alias{select_unique_schemas}
\title{FUNCTION: select_unique_schemas}
\usage{
select_unique_schemas(data)
}
\description{
This function will create a data frame of all of the unique schemas.
Given a input data set, it will return all unique schemas contained in the
original data frame.
}

