% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_x_percent}
\alias{select_x_percent}
\title{FUNCTION: select_x_percent}
\usage{
select_x_percent(d, x)
}
\description{
This function will be used to look at a select percentage of the data provided.
This is referred to as uniform random sampling.
}

