% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_x_percent_across_operators}
\alias{select_x_percent_across_operators}
\title{FUNCTION: select_x_percent_across_operators}
\usage{
select_x_percent_across_operators(d, x)
}
\description{
Select a uniform percentage of data for each operator.
}

