% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_reduced_time}
\alias{summarise_reduced_time}
\title{FUNCTION: summarise_reduced_time}
\usage{
summarise_reduced_time(d)
}
\description{
Calculate the total time (in ms) of performing mutation analysis with the given set of mutants.
}

