% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_average_percent_kept}
\alias{summarize_average_percent_kept}
\title{FUNCTION: summarize_average_percent_kept}
\usage{
summarize_average_percent_kept(d)
}
\description{
Calculate the average percent of mutants kept per-operator across 30 trials.
}

