% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_mean_correlation}
\alias{summarize_mean_correlation}
\title{FUNCTION: summarize_mean_correlation}
\usage{
summarize_mean_correlation(d)
}
\description{
Calculate the mean correlation over 30 trials for a given technique.
}

