% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_mean_cost_reduction}
\alias{summarize_mean_cost_reduction}
\title{FUNCTION: summarize_mean_cost_reduction}
\usage{
summarize_mean_cost_reduction(d)
}
\description{
Calculate the mean cost reduction over 30 trials for a given technique.
}

