% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_operator_percentage_kept}
\alias{summarize_operator_percentage_kept}
\title{FUNCTION: summarize_operator_percentage_kept}
\usage{
summarize_operator_percentage_kept(d)
}
\description{
Calculate the percentage of mutants that are kept for each operator.
}

