% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_operator_time}
\alias{summarize_operator_time}
\title{FUNCTION: summarize_operator_time}
\usage{
summarize_operator_time(d)
}
\description{
Calculate the total time (in ms) of performing mutation analysis with the given set of mutants
on a per-operator basis.
}

