% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_original_time}
\alias{summarize_original_time}
\title{FUNCTION: summarize_original_time}
\usage{
summarize_original_time(d)
}
\description{
Calculate the total time (in ms) of performing mutation analysis with the given set of mutants.
}

