% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_schema_count}
\alias{summarize_schema_count}
\title{FUNCTION: summarize_schema_count}
\usage{
summarize_schema_count(d)
}
\description{
Calculate the total number of mutants per dbms, per schema.
}

