% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_schema_operator_frequencies}
\alias{summarize_schema_operator_frequencies}
\title{FUNCTION: summarize_schema_operator_frequencies}
\usage{
summarize_schema_operator_frequencies(d)
}
\description{
Count the frequencies per operator on a per-schema basis.
}

