% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_mean_correlation_cost_reduction_ratio}
\alias{transform_add_mean_correlation_cost_reduction_ratio}
\title{FUNCTION: transform_add_mean_correlation_cost_reduction_ratio}
\usage{
transform_add_mean_correlation_cost_reduction_ratio(d)
}
\description{
Calculate the correlation, cost reduction ratio (HIB / LIB) metric.
}

