% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_omitted_operators}
\alias{transform_add_omitted_operators}
\title{FUNCTION: transform_add_omitted_operators}
\usage{
transform_add_omitted_operators(d, o, oo)
}
\description{
Add the operators that were omitted from selective random sampling
}

