% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_percent_ignored}
\alias{transform_add_percent_ignored}
\title{FUNCTION: transform_add_percent_ignored}
\usage{
transform_add_percent_ignored(d)
}
\description{
Add the percentage of kept mutants ignored per-operator.
}

