% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_percent_kept}
\alias{transform_add_percent_kept}
\title{FUNCTION: transform_add_percent_kept}
\usage{
transform_add_percent_kept(d)
}
\description{
Add the percentage of kept mutants kept per-operator.
}

