% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_percentage_trial}
\alias{transform_add_percentage_trial}
\title{FUNCTION: transform_add_percentage_trial}
\usage{
transform_add_percentage_trial(d, p, t)
}
\description{
Add the analyzed percentage and current trial to data frame
}

