% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_significance}
\alias{transform_add_significance}
\title{FUNCTION: transform_add_significance}
\usage{
transform_add_significance(d)
}
\description{
Append column with the boolean of whether the result found was significant or not.
}

