% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_start_and_step}
\alias{transform_add_start_and_step}
\title{FUNCTION: transform_add_start_and_step}
\usage{
transform_add_start_and_step(d, s)
}
\description{
Append the randomly-generated start position and appropriate step size
}

