% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_step_size}
\alias{transform_add_step_size}
\title{FUNCTION: transform_add_step_size}
\usage{
transform_add_step_size(d, s)
}
\description{
Append the step size for the hill climbing technique.
}

