% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_add_technique_group}
\alias{transform_add_technique_group}
\title{FUNCTION: transform_add_technique_group}
\usage{
transform_add_technique_group(d)
}
\description{
Add a column to the combined technique data to include the abstract group.
}

