% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_correlation}
\alias{transform_correlation}
\title{FUNCTION: transform_correlation}
\usage{
transform_correlation(d)
}
\description{
Calculate Kendall's tau_b correlation coefficient between the reduced and original mutation score
}

