% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_cost_reduction}
\alias{transform_cost_reduction}
\title{FUNCTION: transform_cost_reduction}
\usage{
transform_cost_reduction(d)
}
\description{
Calculate the reduction in time for performing mutation analysis on the reduced set compared to the original set
}

