% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_error}
\alias{transform_error}
\title{FUNCTION: transform_error}
\usage{
transform_error(d)
}
\description{
Calculate the error --- the difference --- between the original mutation score and the reduced mutation score.
}

