% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_fractional_operator_frequencies}
\alias{transform_fractional_operator_frequencies}
\title{FUNCTION: transform_fractional_operator_frequencies}
\usage{
transform_fractional_operator_frequencies(d)
}
\description{
Calculate the fractional frequencies of each operator per schema.
}

