% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_highest_correlation}
\alias{transform_highest_correlation}
\title{FUNCTION: transform_highest_correlation}
\usage{
transform_highest_correlation(d)
}
\description{
Determine which step is has the highest correlation
}

