% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_keep}
\alias{transform_keep}
\title{FUNCTION: transform_keep}
\usage{
transform_keep(d, all = TRUE)
}
\description{
This column has values representing whether or not a mutant is considered when calculating mutation score
as well as other metrics.
}

