% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_keep_all}
\alias{transform_keep_all}
\title{FUNCTION: transform_keep_all}
\usage{
transform_keep_all(d)
}
\description{
Add a binary value of 1 (for keeping) all mutants for consistency between data tables.
}

