% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_mutant_count}
\alias{transform_mutant_count}
\title{FUNCTION: transform_mutant_count}
\usage{
transform_mutant_count(d)
}
\description{
Count the total number of mutants associated with a schema
}

