% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_original_killed_count}
\alias{transform_original_killed_count}
\title{FUNCTION: transform_original_killed_count}
\usage{
transform_original_killed_count(d)
}
\description{
Count the number of killed mutants
}

