% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_original_mutation_score}
\alias{transform_original_mutation_score}
\title{FUNCTION: transform_original_mutation_score}
\usage{
transform_original_mutation_score(d)
}
\description{
Calculate the mutation score for the original set (number of killed mutants / total number of mutants)
}

