% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_reduced_killed_count}
\alias{transform_reduced_killed_count}
\title{FUNCTION: transform_reduced_killed_count}
\usage{
transform_reduced_killed_count(d)
}
\description{
Count the number of killed mutants
}

