% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_reduced_mutation_score}
\alias{transform_reduced_mutation_score}
\title{FUNCTION: transform_reduced_mutation_score}
\usage{
transform_reduced_mutation_score(d)
}
\description{
Calculate the mutation score for a given set (number of killed mutants / total number of mutants)
}

