% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_reduced_total_count}
\alias{transform_reduced_total_count}
\title{FUNCTION: transform_reduced_total_count}
\usage{
transform_reduced_total_count(d)
}
\description{
Count the total number of mutants for a set
}

