% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_rename_ignore}
\alias{transform_rename_ignore}
\title{FUNCTION: transform_rename_ignore}
\usage{
transform_rename_ignore(d)
}
\description{
Rename column from generic name of 'n' to ignore_count.
}

