% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_replace_correlation}
\alias{transform_replace_correlation}
\title{FUNCTION: transform_replace_correlation}
\usage{
transform_replace_correlation(d)
}
\description{
DESCRIPTION: Make sure that the correlation coeffient calculated by the 'Kendall' R package is renamed
from 'estimate' to 'correlation'
}

