% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_replace_technique}
\alias{transform_replace_technique}
\title{FUNCTION: transform_replace_technique}
\usage{
transform_replace_technique(d)
}
\description{
Update the technique column to include the configuration of the technique also.
This makes comparing all techniques and configurations easier.
}

