% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_update_hill_climbing_applied_operator_data}
\alias{transform_update_hill_climbing_applied_operator_data}
\title{FUNCTION: transform_update_hill_climbing_applied_operator_data}
\usage{
transform_update_hill_climbing_applied_operator_data(d, s)
}
\description{
This function will append the necessary columns to make it match the data from actually performing
the hill climbing technique versus the model.
}

