% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_cost_versus_correlation}
\alias{visualise_cost_versus_correlation}
\title{FUNCTION: visualise_cost_versus_correlation}
\usage{
visualise_cost_versus_correlation(data)
}
\description{
Produces a visualisations comparing reduced cost (time in ms) of the reduced
set of mutants versus the correlation. This visualisation will help determine
whether or not obtaining a higher correlation is worth paying the cost.
}

