% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_cost_versus_mean_absolute_error}
\alias{visualise_cost_versus_mean_absolute_error}
\title{FUNCTION: visualise_cost_versus_mean_absolute_error}
\usage{
visualise_cost_versus_mean_absolute_error(data, s)
}
\description{
Produces a visualisations comparing reduced cost (time in ms) of the reduced
set of mutants versus the mean absolute error at that cost. This visualisation will help determine
whether or not obtaining a lower error is worth paying the cost.
}

