% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_cost_versus_mutation_score}
\alias{visualise_cost_versus_mutation_score}
\title{FUNCTION: visualise_cost_versus_mutation_score}
\usage{
visualise_cost_versus_mutation_score(data, s)
}
\description{
Produces a visualisations comparing reduced cost (time in ms) of the reduced
set of mutants versus the mutation scores. This visualisation will help determine
whether or not obtaining a higher mutation score is worth paying the cost.
}

