% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_k_percent_mutation_score}
\alias{visualise_k_percent_mutation_score}
\title{FUNCTION: visualise_k_percent_mutation_score}
\usage{
visualise_k_percent_mutation_score(data)
}
\description{
Produces a visualisation comparing k_percent and mutation score.
This function will take in a data frame that contains only data for a one schema.
This means that the data will need to be filtered before being passed
to this function. In order to only observe data for one schema first use
the select_individual_schema_data(data, "schema_name")
}

