% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_k_percent_mutation_score_per_schema}
\alias{visualise_k_percent_mutation_score_per_schema}
\title{FUNCTION: visualise_k_percent_mutation_score_per_schema}
\usage{
visualise_k_percent_mutation_score_per_schema(data)
}
\description{
Produces a visualisation comparing k_percent and mutation scores for each schema.
As input it expects all of the data, not just data for a specific schema.
}

