% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_mutation_score_across_schemas}
\alias{visualise_mutation_score_across_schemas}
\title{FUNCTION: visualise_mutation_score_across_schemas}
\usage{
visualise_mutation_score_across_schemas(data, a, b, c, d)
}
\description{
Produces a visualisation of the reduced mutation
scores across all schemas at a specific percentage.
}

