% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_percent_error}
\alias{visualise_percent_error}
\title{FUNCTION: visualise_percent_error}
\usage{
visualise_percent_error(data)
}
\description{
A function displaying distances. This function will be able to display both root mean squared error
and mean absolute error.
}

