% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise_save_graphic}
\alias{visualise_save_graphic}
\title{FUNCTION: visualise_save_graphic}
\usage{
visualise_save_graphic(save_name, save_plot, h, w)
}
\description{
Saves the provided graphic to the provided name.
}

