% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_cost_reduction_all_reduction_technique_groups}
\alias{visualize_cost_reduction_all_reduction_technique_groups}
\title{FUNCTION: visualize_cost_reduction_all_reduction_technique_groups}
\usage{
visualize_cost_reduction_all_reduction_technique_groups(d)
}
\description{
Summary graphic comparing the aggregated techniques cost reduction.
}

