% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_hill_climbing_cost_reduction_all_dbms}
\alias{visualize_hill_climbing_cost_reduction_all_dbms}
\title{FUNCTION: visualize_hill_climbing_cost_reduction_all_dbms}
\usage{
visualize_hill_climbing_cost_reduction_all_dbms(d)
}
\description{
Visualize the cost reduction for all of the DBMSs when comparing the operator models.
}

