% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_mean_corr_cost_reduction_head_to_head}
\alias{visualize_mean_corr_cost_reduction_head_to_head}
\title{FUNCTION: visualize_mean_corr_cost_reduction_head_to_head}
\usage{
visualize_mean_corr_cost_reduction_head_to_head(d)
}
\description{
Dot plot comparing the mean correlation and cost reduction for two techniques.
}

