% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_operator_sampling_mutation_scores}
\alias{visualize_operator_sampling_mutation_scores}
\title{FUNCTION: visualize_operator_sampling_mutation_scores}
\usage{
visualize_operator_sampling_mutation_scores(d, trans = FALSE)
}
\description{
Visualize mutation scores for the operator sampling reduction technique
}

