% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_original_mutation_score_per_dbms}
\alias{visualize_original_mutation_score_per_dbms}
\title{FUNCTION: visualize_original_mutation_score_per_dbms}
\usage{
visualize_original_mutation_score_per_dbms(d)
}
\description{
Summary graphic showing the original mutation scores per dbms.
}

