% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_cost_reduction}
\alias{visualize_plot_cost_reduction}
\title{FUNCTION: visualize_plot_cost_reduction}
\usage{
visualize_plot_cost_reduction(d)
}
\description{
Produces a visualization of the cost reduction from the original and reduced mutant sets across
schemas for thirty trials.
}

