% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_cost_reduction_all_reduction_techniques_box}
\alias{visualize_plot_cost_reduction_all_reduction_techniques_box}
\title{FUNCTION: visualize_plot_cost_reduction_all_reduction_techniques_box}
\usage{
visualize_plot_cost_reduction_all_reduction_techniques_box(d)
}
\description{
Plot fractional cost reduction for each technique.
}

