% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_cost_reduction_multiple_dbms}
\alias{visualize_plot_cost_reduction_multiple_dbms}
\title{FUNCTION: visualize_plot_cost_reduction_mulitple_dbms}
\usage{
visualize_plot_cost_reduction_multiple_dbms(d)
}
\description{
Produces a visualization of the cost reduction values of the operator model applied to multiple DBMSs.
}

