% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_effectsize}
\alias{visualize_plot_effectsize}
\title{FUNCTION: visualize_plot_effectsize}
\usage{
visualize_plot_effectsize(d)
}
\description{
Plot effect size of all pairs of reduction techniques.
}

