% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_effectsize_pres}
\alias{visualize_plot_effectsize_pres}
\title{FUNCTION: visualize_plot_effectsize_pres}
\usage{
visualize_plot_effectsize_pres(d)
}
\description{
Plot effect size of all pairs of reduction techniques.
}

