% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_error}
\alias{visualize_plot_error}
\title{FUNCTION: visualize_plot_error}
\usage{
visualize_plot_error(d)
}
\description{
Produces a visualization of the error --- the difference --- between the original and reduced mutation
scores.
}

