% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_fractional_operator_frequencies}
\alias{visualize_plot_fractional_operator_frequencies}
\title{FUNCTION: visualize_plot_fractional_operator_frequencies}
\usage{
visualize_plot_fractional_operator_frequencies(d)
}
\description{
Produces a visualization of the frequencies of mutants per operator.
}

