% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_mae}
\alias{visualize_plot_mae}
\title{FUNCTION: visualize_plot_mae}
\usage{
visualize_plot_mae(d)
}
\description{
Produces a visualization of the MAE --- the mean absolute error --- between the original and reduced mutation
scores across thirty trials
}

