% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_percentage_cost}
\alias{visualize_plot_percentage_cost}
\title{FUNCTION: visualize_plot_percentage_cost}
\usage{
visualize_plot_percentage_cost(d)
}
\description{
Produces a visualization of the time to evaluate the mutants chosen for the reduced set.
}

