% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_ratio_pres}
\alias{visualize_plot_ratio_pres}
\title{FUNCTION: visualize_plot_ratio_pres}
\usage{
visualize_plot_ratio_pres(d)
}
\description{
Produces a box plot comparing ratio values
}

