% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_rmse}
\alias{visualize_plot_rmse}
\title{FUNCTION: visualize_plot_rmse}
\usage{
visualize_plot_rmse(d)
}
\description{
Produces a visualization of the RMSE --- the root mean squared error --- between the original and reduced mutation
scores across thirty trials
}

