% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_wilcoxon}
\alias{visualize_plot_wilcoxon}
\title{FUNCTION: visualize_plot_wilcoxon}
\usage{
visualize_plot_wilcoxon(d)
}
\description{
Plot ranked sum of all pairs of reduction techniques.
}

