% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_plot_wilcoxon_pres}
\alias{visualize_plot_wilcoxon_pres}
\title{FUNCTION: visualize_plot_wilcoxon_pres}
\usage{
visualize_plot_wilcoxon_pres(d)
}
\description{
Plot ranked sum of all pairs of reduction techniques.
}

