% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_ratio_head_to_head}
\alias{visualize_ratio_head_to_head}
\title{FUNCTION: visualize_corr_cost_reduction_head_to_head}
\usage{
visualize_ratio_head_to_head(d)
}
\description{
Box plot comparing the correlation and cost reduction for two techniques.
}

