% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_save_graphic}
\alias{visualize_save_graphic}
\title{FUNCTION: visualize_save_graphic}
\usage{
visualize_save_graphic(save_name, save_plot, w, h)
}
\description{
Saves the provided graphic to the provided name.
}

