% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_across_operators}
\alias{analyse_across_operators}
\title{FUNCTION: analyse_across_operators}
\usage{
analyse_across_operators(d)
}
\description{
This function will perform sampling across all operators --- this is not operators selection as alll
operators will be present.
}

