% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_calculations}
\alias{analyse_calculations}
\title{FUNCTION: analyse_calculations}
\usage{
analyse_calculations(d)
}
\description{
Calculate the effectiveness of a reduction technique.
The metrics to determine effectiveness will be Kendall's tau_b correlation coefficient,
mae and rmse.
}

