% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_correlation}
\alias{analyse_correlation}
\title{FUNCTION: analyse_correlation}
\usage{
analyse_correlation(data)
}
\description{
This function for each percent will calculate kendall's tau_b correlation between the given percent's
mutation score and the mutation score of 100 percent of the mutants for all schemas at that percentage.
The reason it is across all schemas is due to the fact that for one schema the original mutation score
column contains fixed values and the correlation function expects two vectors with non-fixed values.
}

