% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_keep_calculations}
\alias{analyze_keep_calculations}
\title{FUNCTION: analyze_keep_calculations}
\usage{
analyze_keep_calculations(d, trial)
}
\description{
Calculate the effectiveness of a reduction technique on a per-trial, per-schema.
}

