% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze_percent_calculations}
\alias{analyze_percent_calculations}
\title{FUNCTION: analyze_percent_calculations}
\usage{
analyze_percent_calculations(d)
}
\description{
Calculate the effectiveness of a reduction technique across schemas.
The metrics to determine effectiveness will be Kendall's tau_b correlation coefficient,
mae and rmse.
}

