% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{FUNCTION: calculate_correlation}
\usage{
calculate_correlation(x, y)
}
\description{
This function will calculate the correlation between the reduced and the original
mutation score for a given percent. This is a helper function for the transform_correlation function
}

