% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{random_sampling}
\alias{random_sampling}
\title{FUNCTION: random_sampling}
\usage{
random_sampling(data)
}
\description{
This is the function that will perform random sampling. There is a function
called analyse_random_sampling which utilizes this function to perform
random sampling in parallel.
}

