% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_keep}
\alias{transform_keep}
\title{FUNCTION: transform_keep}
\usage{
transform_keep(d)
}
\description{
Add a binary value of 1 (for keeping) and 0 (for not keeping) a mutant.
This will be used for hill climbing and other learning approaches.
}

